move sorting stuff to models.sorting
