+++ WalletWasabi.Gui/Controls/SortingArrow.cs
@@ -12,33 +12,12 @@ using Avalonia.Styling;
 using ReactiveUI;
 using System;
 using System.ComponentModel;
-using System.Diagnostics.CodeAnalysis;
 using System.Reactive.Linq;
 using System.Runtime.CompilerServices;
+using WalletWasabi.Gui.Models.Sorting;
 
 namespace WalletWasabi.Gui.Controls
 {
-	public enum SortOrder
-	{
-		None,
-		Increasing,
-		Decreasing
-	}
-	
-	public struct SortingPreference : IEquatable<SortingPreference>
-	{
-		public SortingPreference(SortOrder sortOrder, string colTarget)
-		{
-			SortOrder = sortOrder;
-			ColumnTarget = colTarget;
-		}
-
-		public SortOrder SortOrder { get; set; }
-		public string ColumnTarget { get; set; }
-
-		public bool Equals([AllowNull] SortingPreference other) => SortOrder == other.SortOrder && ColumnTarget == other.ColumnTarget;
-	}
-
 	public class SortingArrow : Button, IStyleable
 	{
 		public static readonly StyledProperty<SortOrder> SortDirectionProperty =
