+++ WalletWasabi.Gui/Models/Sorting/SortOrder.cs
@@ -1,0 +1,9 @@
+namespace WalletWasabi.Gui.Models.Sorting
+{
+	public enum SortOrder
+	{
+		None,
+		Increasing,
+		Decreasing
+	}
+}
